refactor(cycles): move the policy read model out of the command module (#783)
Folds two findings from the cross-family maintainability review.

The command module had absorbed the HTTP read model: 184 of its 194 new
lines were effective-policy hydration and provenance mapping, and
`EffectiveCyclePolicy` had grown two hidden modes — lean for preview,
apply and conflict, rich for reads — told apart only by default-empty
metadata. A caller could receive a half-populated object with nothing in
the type to say so. The read model now lives in
`behavior_policy_read_model.py` with an explicit rich result type, and
`behavior_policy.py` drops 189 lines back to command-domain work.

Apply and revert/apply also duplicated conflict handling, result checks,
the effective-policy reload, response assembly and commit. Both now call
one `_finalize_policy_apply_result`.

No HTTP contract change: same routes, same status codes, same bodies.
`tests/test_cycle_behavior_policy_api.py` passes unmodified.

The remaining blocking finding — 11 synchronized edits to add one policy
field — is filed as #795 and gates slice 3, because its remedy renames
API fields and that is a contract decision.
